# frontend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/scorecard/mock/preview": {
      "post": {
        "description": "Generates a mock scorecard for a given GitHub username and returns a frontend-ready UI model. Requires internal API key.",
        "operationId": "ScorecardController_preview",
        "parameters": [
          {
            "name": "X-Internal-Key",
            "in": "header",
            "description": "Internal API key (required)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewScorecardRequestDto"
              },
              "examples": {
                "example1": {
                  "summary": "Basic request",
                  "value": {
                    "githubUsername": "octocat"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully generated UI scorecard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardUiDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid internal API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Preview scorecard (UI model)",
        "tags": [
          "Scorecard"
        ]
      }
    },
    "/api/scorecard/mock/preview/raw": {
      "post": {
        "description": "Returns full raw scoring data for debugging and internal analysis. Not intended for frontend usage.",
        "operationId": "ScorecardController_previewRaw",
        "parameters": [
          {
            "name": "X-Internal-Key",
            "in": "header",
            "description": "Internal API key (required)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewScorecardRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Raw scorecard data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardRawResponseDto"
                }
              }
            }
          }
        },
        "summary": "Preview scorecard (raw model)",
        "tags": [
          "Scorecard"
        ]
      }
    },
    "/api/scorecard/me": {
      "get": {
        "description": "Returns the authenticated user’s scorecard formatted for frontend display.",
        "operationId": "ScorecardController_getMyScorecard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User scorecard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardUiDto"
                }
              }
            }
          },
          "404": {
            "description": "No scorecard found. User must trigger GitHub sync first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my scorecard (UI)",
        "tags": [
          "Scorecard"
        ]
      }
    },
...
    "/api/scorecard/me/raw": {
      "get": {
        "description": "Returns raw internal scorecard data for the authenticated user.",
        "operationId": "ScorecardController_getMyScorecardRaw",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Raw scorecard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardRawResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my scorecard (raw)",
        "tags": [
          "Scorecard"
        ]
      }
    },
    "/api/scorecard/{username}/raw": {
      "get": {
        "description": "Returns raw cached scorecard for debugging or internal usage.",
        "operationId": "ScorecardController_getPublicScorecardRaw",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "example": "octocat",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw scorecard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardRawResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public scorecard (raw)",
        "tags": [
          "Scorecard"
        ]
      }
    },
...
    "/sync/wallet/challenge": {
      "get": {
        "description": "Creates a time-limited cryptographic challenge used to verify wallet ownership before linking a Solana wallet to the user account. Flow [server gives you a challenge-message >>> wallet signs challenge-message >>> server verifies signature match for challenge-message] This is step 1 in the flow ",
        "operationId": "WalletSyncController_getChallenge",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Challenge successfully generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeResponseDto"
                },
                "example": {
                  "challenge": "link-wallet:user123:1713940000000:a1b2c3"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletSyncErrorResponseDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Unauthorized",
                  "error": "Unauthorized"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate wallet linking challenge",
        "tags": [
          "Wallet Sync"
        ]
      }
    },
    "/sync/wallet": {
      "post": {
        "description": "Verifies a Solana wallet signature against a previously generated challenge and links the wallet to the authenticated user account.",
        "operationId": "WalletSyncController_linkWallet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Wallet address and cryptographic signature",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkWalletRequestDto"
              },
              "examples": {
                "example1": {
                  "summary": "Valid request example",
                  "value": {
                    "walletAddress": "7Gg3...SolanaAddressExample...",
                    "signature": "5hK9...Base58Signature..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Wallet successfully linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkWalletResponseDto"
                },
                "example": {
                  "linked": true,
                  "solanaAddress": "7Gg3...SolanaAddressExample..."
                }
              }
            }
          },
          "400": {
            "description": "Invalid wallet address format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletSyncErrorResponseDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Invalid Solana wallet address",
                  "error": "Bad Request"
                }
              }
            }
          },
          "401": {
            "description": "Signature verification failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletSyncErrorResponseDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "Wallet signature invalid",
                  "error": "Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Challenge expired or candidate profile missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletSyncErrorResponseDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Challenge expired or not found",
                  "error": "Not Found"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify wallet signature and link wallet",
        "tags": [
          "Wallet Sync"
        ]
      }
    },
...
  },
  "info": {
    "title": "a16zero Employer API",
    "description": "Backend MVP for Employer platform features and Account Abstraction auth verification.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "internal-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-key",
        "description": "Internal API key"
      }
    },
    "schemas": {
...
      "PreviewScorecardRequestDto": {
        "type": "object",
        "properties": {}
      },
      "ScorecardUiDto": {
        "type": "object",
        "properties": {
          "profile": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "description": "GitHub username of the candidate"
              },
              "avatarUrl": {
                "type": "string",
                "format": "uri"
              },
              "primaryCohort": {
                "type": "string"
              },
              "seniority": {
                "type": "string",
                "enum": [
                  "JUNIOR",
                  "MID",
                  "SENIOR",
                  "LEAD"
                ]
              },
              "summary": {
                "type": "string"
              }
            },
            "required": [
              "username",
              "primaryCohort",
              "seniority",
              "summary"
            ]
          },
          "score": {
            "type": "object",
            "properties": {
              "value": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "percentile": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "isWithheld": {
                "type": "object",
                "properties": {
                  "value": {
                    "type": "boolean"
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "value"
                ]
              }
            },
            "required": [
              "value",
              "percentile",
              "isWithheld"
            ]
          },
          "trust": {
            "type": "object",
            "properties": {
              "level": {
                "type": "string"
              },
              "risk": {
                "type": "string",
                "enum": [
                  "LOW_RISK",
                  "MEDIUM_RISK",
                  "HIGH_RISK",
                  "INSUFFICIENT_DATA"
                ]
              },
              "label": {
                "type": "string"
              },
              "guidance": {
                "type": "string"
              }
            },
            "required": [
              "level",
              "risk",
              "label",
              "guidance"
            ]
          },
          "insights": {
            "type": "object",
            "properties": {
              "capabilities": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "enum": [
                        "backend",
                        "frontend",
                        "devops"
                      ]
                    },
                    "label": {
                      "type": "string"
                    },
                    "score": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 1
                    },
                    "displayScore": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 100
                    },
                    "confidence": {
                      "type": "string"
                    },
                    "strength": {
                      "type": "string",
                      "enum": [
                        "strong",
                        "moderate",
                        "weak"
                      ]
                    }
                  },
                  "required": [
                    "key",
                    "label",
                    "score",
                    "displayScore",
                    "confidence",
                    "strength"
                  ]
                }
              },
              "highlights": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "gaps": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "caveats": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "ownership": {
                "type": "object",
                "properties": {
                  "ownedProjects": {
                    "type": "number"
                  },
                  "activelyMaintained": {
                    "type": "number"
                  },
                  "confidence": {
                    "type": "string"
                  }
                },
                "required": [
                  "ownedProjects",
                  "activelyMaintained",
                  "confidence"
                ]
              },
              "impact": {
                "type": "object",
                "properties": {
                  "activityLevel": {
                    "type": "string"
                  },
                  "consistency": {
                    "type": "string"
                  },
                  "externalContributions": {
                    "type": "number"
                  },
                  "confidence": {
                    "type": "string"
                  }
                },
                "required": [
                  "activityLevel",
                  "consistency",
                  "externalContributions",
                  "confidence"
                ]
              },
              "reputation": {
                "nullable": true
              },
              "privateWorkNote": {
                "type": "string"
              },
              "organizations": {
                "type": "array",
                "items": {}
              },
              "interactionProfile": {
                "nullable": true
              },
              "stack": {
                "type": "object",
                "properties": {
                  "languages": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "tools": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "languages",
                  "tools"
                ]
              },
              "web3": {
                "nullable": true
              }
            },
            "required": [
              "capabilities",
              "highlights",
              "gaps",
              "caveats",
              "ownership",
              "impact"
            ]
          }
        },
        "required": [
          "profile",
          "score",
          "trust",
          "insights"
        ]
      },
      "ScorecardErrorResponseDto": {
        "type": "object",
        "properties": {}
      },
      "ScorecardRawResponseDto": {},
...
      "ChallengeResponseDto": {
        "type": "object",
        "properties": {
          "challenge": {
            "type": "string",
            "description": "Cryptographic challenge string used to verify wallet ownership",
            "example": "link-wallet:user123:1713940000000:a1b2c3"
          }
        },
        "required": [
          "challenge"
        ]
      },
      "WalletSyncErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "object",
            "example": "Validation failed"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "LinkWalletRequestDto": {
        "type": "object",
        "properties": {
          "walletAddress": {
            "type": "string",
            "description": "Solana wallet public address (Base58 encoded)",
            "example": "7Gg3xYzExampleSolanaWalletAddress123"
          },
          "signature": {
            "type": "string",
            "description": "Base58 encoded signature of the challenge signed by the wallet private key",
            "example": "5hK9xExampleSignatureBase58EncodedString"
          }
        },
        "required": [
          "walletAddress",
          "signature"
        ]
      },
      "LinkWalletResponseDto": {
        "type": "object",
        "properties": {
          "linked": {
            "type": "boolean",
            "description": "Indicates whether wallet was successfully linked",
            "example": true
          },
          "solanaAddress": {
            "type": "string",
            "description": "Linked Solana wallet address",
            "example": "7Gg3xYzExampleSolanaWalletAddress123"
          }
        },
        "required": [
          "linked",
          "solanaAddress"
        ]
      },
...
    }
  }
}